+++ packages/lila-webpack-config/src/rules.js
@@ -1,11 +1,6 @@
 import autoprefixer from 'autoprefixer';
 import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-import { baseType, reactType, vueType, reactVueType } from './data';
-
-const vueLoader = () => ({
-  loader: 'vue-loader',
-  test: /\.vue$/,
-});
+import { baseType, reactType, reactVueType } from './data';
 
 const baseBabelLoader = ({
   babelImport,
@@ -113,12 +108,13 @@ export const babelLoader = ({
     );
   }
 
-  if (makeType === vueType || makeType === reactVueType) {
-    rules.push(vueLoader());
-  }
-
   return rules;
 };
+
+export const vueLoader = () => ({
+  loader: 'vue-loader',
+  test: /\.vue$/,
+});
 
 export const urlLoader = ({ extensions }) => ({
   loader: 'url-loader',