+++ packages/lila-webpack-lib-config/src/data.js
@@ -1,0 +1,3 @@
+export const baseType = 'base';
+export const reactType = 'react';
+export const vueType = 'vue';