+++ packages/lila-webpack-lib-config/src/start.js
@@ -4,13 +4,13 @@ import base from './base';
 
 const { join } = path;
 
-export default ({ lila, webpack, entry, cmd, config }) => {
+export default ({ lila, webpack, entry, cmd, config, makeType }) => {
   const { HotModuleReplacementPlugin } = webpack;
   const { getSettings } = lila;
   const [root, devDir] = getSettings(['root', 'dev']);
   const devPath = join(root, devDir);
 
-  const baseConfig = base({ lila, webpack, entry, cmd, config });
+  const baseConfig = base({ lila, webpack, entry, cmd, config, makeType });
 
   baseConfig.plugins.unshift(new HotModuleReplacementPlugin());
   baseConfig.plugins.push(